+++ packages/@runlightyear/cli/src/commands/dev/index.ts
@@ -1,17 +1,12 @@
-import { Command } from "commander";
+import { Command, Option } from "commander";
 import getPusher from "../../shared/getPusher";
 import getPusherCredentials from "../../shared/getPusherCredentials";
 import handleRunLocal from "./handleRunLocal";
 import nodemon from "nodemon";
-import execDeploy from "../../shared/execDeploy";
-import execUnsubscribeAfterDeploy from "../../shared/execUnsubscribeAfterDeploy";
-import execSubscribeAfterDeploy from "../../shared/execSubscribeAfterDeploy";
 import { terminal } from "terminal-kit";
 import { setLogDisplayLevel } from "../../shared/setLogDisplayLevel";
 import { prepareConsole } from "../../logging";
 import handleResubscribe from "./handleResubscribe";
-import execSubscribeProps from "../../shared/execSubscribeProps";
-import createDeploy from "../../shared/createDeploy";
 import execDeployAndSubscribe from "../../shared/execDeployAndSubscribe";
 
 export const dev = new Command("dev");
@@ -25,8 +20,11 @@ const largeLogo = `
       |___/           |___/                 
 `;
 
+let firstDeploy = true;
+
 dev
   .description("Automatically compile and deploy on changes to source")
+  .addOption(new Option("--dev").hideHelp())
   .action(async () => {
     terminal(largeLogo);
 
@@ -81,6 +79,13 @@ dev
 
     nodemon.on("exit", async () => {
       await execDeployAndSubscribe();
+
+      if (firstDeploy) {
+        terminal(
+          "\n\nDashboard is available at: https://app.runlightyear.com\n"
+        );
+        firstDeploy = false;
+      }
 
       terminal("\n\nWaiting for file changes...\n");
       terminal("press h for help, press q to quit\n");