+++ packages/@runlightyear/cli/src/shared/createDeploy.ts
@@ -1,4 +1,5 @@
 import fetch from "node-fetch";
+import { getApiKey, getBaseUrl, getEnvName } from "@runlightyear/lightyear";
 
 export interface CreateDeployProps {
   compiledCode: Buffer;
@@ -9,9 +10,9 @@ export default async function createDeploy(
 ): Promise<string> {
   const { compiledCode } = props;
 
-  const baseUrl = process.env.BASE_URL;
-  const envName = process.env.ENV_NAME;
-  const apiKey = process.env.API_KEY;
+  const baseUrl = getBaseUrl();
+  const envName = getEnvName();
+  const apiKey = getApiKey();
 
   let response;