+++ packages/@runlightyear/cli/src/shared/uploadSubscribeResult.ts
@@ -1,5 +1,6 @@
 import fetch from "node-fetch";
 import { terminal } from "terminal-kit";
+import { getApiKey, getBaseUrl, getEnvName } from "@runlightyear/lightyear";
 
 export interface UploadSubscribeResultProps {
   webhookName: string;
@@ -13,9 +14,9 @@ export default async function uploadSubscribeResult(
 ) {
   const { webhookName, status, unsubscribeProps, logs } = props;
 
-  const baseUrl = process.env.BASE_URL;
-  const envName = process.env.ENV_NAME;
-  const apiKey = process.env.API_KEY;
+  const baseUrl = getBaseUrl();
+  const envName = getEnvName();
+  const apiKey = getApiKey();
 
   const activityResponse = await fetch(
     `${baseUrl}/api/v1/envs/${envName}/webhooks/${webhookName}/subscription/activities`,