+++ packages/@runlightyear/lightyear/src/handler/index.ts
@@ -11,6 +11,9 @@ import { handlerResult } from "./handlerResult";
 import { prefixedRedactedConsole } from "../logging";
 import { prepareConsole } from "../logging/prepareConsole";
 import { handleSubscribeProps } from "./handleSubscribeProps";
+import { getEnvName } from "../util/getEnvName";
+import { getApiKey } from "../util/getApiKey";
+import { getBaseUrl } from "../util/getBaseUrl";
 
 interface RepoInvocation extends APIGatewayEvent {
   operation: string;
@@ -31,22 +34,22 @@ export const handler = async (
   prefixedRedactedConsole.initialize();
   console.debug("Initialized console");
 
-  const envName = process.env.ENV_NAME;
+  const envName = getEnvName();
   if (!envName) {
     console.error("Environment variable ENV_NAME not set");
     return handlerResult(400, "Environment variable ENV_NAME not set");
   } else {
-    console.debug("In environment", process.env.ENV_NAME);
+    console.debug("In environment", envName);
   }
 
-  const apiKey = process.env.API_KEY;
+  const apiKey = getApiKey();
   if (apiKey) {
     prefixedRedactedConsole.addSecrets([apiKey]);
   } else {
     return handlerResult(400, "Environment variable API_KEY not set");
   }
 
-  if (!process.env.BASE_URL) {
+  if (!getBaseUrl()) {
     return handlerResult(400, "Environment variable BASE_URL not set");
   }