+++ packages/@runlightyear/lightyear/src/index.ts
@@ -48,6 +48,9 @@ import { handler } from "./handler";
 import toTimestamp from "./util/toTimestamp";
 import { PollingData, WebhookDeliveryData } from "./base/runData";
 import deCamelize from "./util/deCamelize";
+import { getApiKey } from "./util/getApiKey";
+import { getBaseUrl } from "./util/getBaseUrl";
+import { getEnvName } from "./util/getEnvName";
 
 const SKIPPED = "SKIPPED";
 
@@ -69,6 +72,9 @@ export {
   deCamelize,
   HttpProxyResponseError,
   SKIPPED,
+  getApiKey,
+  getBaseUrl,
+  getEnvName,
 };
 
 export type {