+++ packages/@runlightyear/lightyear/src/util/getApiKey.ts
@@ -1,0 +1,3 @@
+export function getApiKey() {
+  return process.env.LIGHTYEAR_API_KEY || process.env.API_KEY;
+}