+++ packages/@runlightyear/lightyear/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @runlightyear/lightyear
 
+## 0.15.0
+
+### Minor Changes
+
+- 36fa04b: Support LIGHTYEAR_API_KEY environment variable
+
 ## 0.14.0
 
 ### Minor Changes